Split off "required" packages into their own apk get line
The reason we do this is because if we need future runtime dependencies
that one of the build-deps pulls in, it won't get removed during the
apk del phase (because we've explicitly asked for it to be installed).

Example:

apk add libpq
apk add --virtual postgresql-dev
apk del --purge build-deps

postgresql-dev pulls in postgresql.  One of those two pulls in libpq
as a dependency.  Thus, if we had not explicitly asked for it in the
first "apk add libpq" line, it would get removed during "apk del".

We also use --virtual here to make our life a little easier.